requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
amqp==1.4.9               # via -r requirements/base.txt, kombu
anyjson==0.3.3            # via -r requirements/base.txt, kombu
astroid==1.6.6            # via pylint, pylint-celery
backports.functools-lru-cache==1.6.1  # via astroid, isort, pylint
billiard==3.3.0.23        # via -r requirements/base.txt, celery
click-log==0.3.2          # via edx-lint
click==7.1.2              # via click-log, edx-lint
configparser==4.0.2       # via pylint
edx-lint==1.4.1           # via -c requirements/constraints.txt, -r requirements/test.in
edx-opaque-keys[django]==0.4.4  # via -c requirements/constraints.txt, -r requirements/base.txt
enum34==1.1.10            # via astroid
future==0.18.2 ; python_version < "3.0"  # via -r requirements/base.txt
futures==3.3.0            # via isort
isort==4.3.21             # via pylint
kombu==3.0.37             # via -r requirements/base.txt, celery
lazy-object-proxy==1.5.1  # via astroid
mccabe==0.6.1             # via pylint
pbr==5.5.1                # via -r requirements/base.txt, stevedore
pycodestyle==2.6.0        # via -r requirements/test.in
pylint-celery==0.3        # via edx-lint
pylint-django==0.11.1     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==1.9.5             # via -r requirements/test.in, edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.11.0           # via -r requirements/base.txt, edx-opaque-keys
pytz==2020.4              # via -r requirements/base.txt, celery, django
singledispatch==3.4.0.3   # via astroid, pylint
six==1.15.0               # via -r requirements/base.txt, astroid, edx-lint, edx-opaque-keys, pylint, singledispatch, stevedore
stevedore==1.32.0         # via -r requirements/base.txt, edx-opaque-keys
web-fragments==0.3.2      # via -r requirements/base.txt
wrapt==1.12.1             # via astroid
